Refactor Processors
