+++ src/Sylius/Bundle/ApiBundle/Resources/config/api_platform/resources/Country.xml
@@ -16,7 +16,7 @@
     xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
     xsi:schemaLocation="https://api-platform.com/schema/metadata/resources-3.0 https://api-platform.com/schema/metadata/resources-3.0.xsd"
 >
-    <resource class="Sylius\Component\Addressing\Model\Country" processor="Sylius\Bundle\ApiBundle\StateProcessor\CountryProcessor">
+    <resource class="Sylius\Component\Addressing\Model\Country" processor="sylius.api.state_processor.admin.country.persist">
         <validationContext>
             <values>
                 <value name="groups">
