+++ src/Sylius/Bundle/ApiBundle/Resources/config/api_platform/resources/Product.xml
@@ -271,7 +271,7 @@
                 </normalizationContext>
             </operation>
 
-            <operation class="ApiPlatform\Metadata\Delete" uriTemplate="/admin/products/{code}" processor="sylius_api.state_processor.delete.product"/>
+            <operation class="ApiPlatform\Metadata\Delete" uriTemplate="/admin/products/{code}" processor="sylius_api.state_processor.admin.product.remove"/>
 
         </operations>
     </resource>
