+++ src/Sylius/Bundle/ApiBundle/StateProcessor/Admin/Product/RemoveProcessor.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\ApiBundle\StateProcessor\Delete;
+namespace Sylius\Bundle\ApiBundle\StateProcessor\Admin\Product;
 
 use ApiPlatform\Metadata\DeleteOperationInterface;
 use ApiPlatform\Metadata\Operation;
@@ -21,7 +21,7 @@ use Sylius\Bundle\ApiBundle\Exception\ProductCannotBeRemoved;
 use Sylius\Component\Core\Model\ProductInterface;
 use Webmozart\Assert\Assert;
 
-final readonly class ProductProcessor implements ProcessorInterface
+final readonly class RemoveProcessor implements ProcessorInterface
 {
     public function __construct(
         private ProcessorInterface $removeProcessor,
@@ -31,13 +31,13 @@ final readonly class ProductProcessor implements ProcessorInterface
     /**
      * @param ProductInterface $data
      */
-    public function process(mixed $data, Operation $operation, array $uriVariables = [], array $context = [])
+    public function process(mixed $data, Operation $operation, array $uriVariables = [], array $context = []): void
     {
         Assert::isInstanceOf($data, ProductInterface::class);
         Assert::isInstanceOf($operation, DeleteOperationInterface::class);
 
         try {
-            return $this->removeProcessor->process($data, $operation, $uriVariables, $context);
+            $this->removeProcessor->process($data, $operation, $uriVariables, $context);
         } catch (ForeignKeyConstraintViolationException) {
             throw new ProductCannotBeRemoved();
         }
